requirements.txt
# pip freeze
python-dotenv==0.21.0
requests==2.28.1
beautifulsoup4==4.11.2
nltk==3.8.1
langdetect==1.0.9
numpy==1.24.2
scikit-learn==1.2.1
caribou==0.3.0
python-telegram-bot==20.1
# 